Fixed bug in JS interaction:
Tool can now properly load and display data of the automated entities.
Further extension of functionality of the JS is still required. Need
extra way of adding other fields AND allow the annotation to be linked
to an extra/existing annotation (treat Annotation nodes as if there's
a new annotation highlighted)